Refactor: window 관련 함수 test-utils로 분리
